+++ src/main/java/net/frozenblock/lib/cape/impl/Cape.java
@@ -20,8 +20,61 @@ package net.frozenblock.lib.cape.impl;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import net.fabricmc.fabric.api.attachment.v1.AttachmentRegistry;
+import net.fabricmc.fabric.api.attachment.v1.AttachmentSyncPredicate;
+import net.fabricmc.fabric.api.attachment.v1.AttachmentType;
+import net.frozenblock.lib.FrozenLibConstants;
+import net.frozenblock.lib.cape.api.CapeUtil;
+import net.minecraft.core.ClientAsset;
+import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.codec.ByteBufCodecs;
+import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.resources.Identifier;
 
-public record Cape(Identifier id, Component name, Identifier texture, Optional<List<UUID>> allowedPlayers) {
+public record Cape(Identifier id, Component name, CapeTexture texture, Optional<List<UUID>> allowedPlayers) {
+	public static final StreamCodec<FriendlyByteBuf, Optional<Cape>> NETWORK_CODEC = StreamCodec.of(Cape::writeToStream, Cape::createFromStream);
+	public static final AttachmentType<Optional<Cape>> ATTACHMENT_TYPE = AttachmentRegistry.create(
+		FrozenLibConstants.id("cape"),
+		builder -> {
+			builder.syncWith(NETWORK_CODEC, AttachmentSyncPredicate.all());
+			builder.copyOnDeath();
+		}
+	);
+
+	public Cape(Identifier id, Component name, Identifier texture, Optional<List<UUID>> allowedPlayers) {
+		this(id, name, new CapeTexture(texture), allowedPlayers);
+	}
+
+	public static void init() {}
+
+	public boolean dummy() {
+		return this.id.getPath().equals("dummy");
+	}
+
+	private static void writeToStream(FriendlyByteBuf output, Optional<Cape> cape) {
+		ByteBufCodecs.optional(Identifier.STREAM_CODEC).encode(output, cape.map(Cape::id));
+	}
+
+	public static Optional<Cape> createFromStream(FriendlyByteBuf input) {
+		return ByteBufCodecs.optional(Identifier.STREAM_CODEC).decode(input).flatMap(CapeUtil::getCape);
+	}
+
+	public static class CapeTexture implements ClientAsset.Texture {
+		private final Identifier texture;
+
+		private CapeTexture(Identifier texture) {
+			this.texture = texture;
+		}
+
+		@Override
+		public Identifier texturePath() {
+			return this.texture;
+		}
+
+		@Override
+		public Identifier id() {
+			return this.texture;
+		}
+	}
 }