+++ src/main/java/net/frozenblock/lib/cape/mixin/client/CapeLayerMixin.java
@@ -18,39 +18,19 @@
 package net.frozenblock.lib.cape.mixin.client;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import com.llamalad7.mixinextras.sugar.Share;
-import com.llamalad7.mixinextras.sugar.ref.LocalRef;
-import com.mojang.blaze3d.vertex.PoseStack;
+import com.llamalad7.mixinextras.sugar.Local;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.frozenblock.lib.cape.client.impl.AvatarCapeInterface;
-import net.minecraft.client.renderer.SubmitNodeCollector;
+import net.frozenblock.lib.cape.client.api.ClientCapeUtil;
 import net.minecraft.client.renderer.entity.layers.CapeLayer;
 import net.minecraft.client.renderer.entity.state.AvatarRenderState;
 import net.minecraft.core.ClientAsset;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Environment(EnvType.CLIENT)
 @Mixin(CapeLayer.class)
 public class CapeLayerMixin {
-
-	@Inject(
-		method = "submit(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;ILnet/minecraft/client/renderer/entity/state/AvatarRenderState;FF)V",
-		at = @At(
-			value = "FIELD",
-			target = "Lnet/minecraft/client/renderer/entity/state/AvatarRenderState;skin:Lnet/minecraft/world/entity/player/PlayerSkin;"
-		)
-	)
-	private static void frozenLib$captureFrozenLibCape(
-		PoseStack poseStack, SubmitNodeCollector collector, int i, AvatarRenderState renderState, float f, float g, CallbackInfo info,
-		@Share("frozenLib$newCapeTexture") LocalRef<ClientAsset.Texture> newCapeAssetRef
-	) {
-		if (!(renderState instanceof AvatarCapeInterface capeInterface)) return;
-		newCapeAssetRef.set(capeInterface.frozenLib$getCape());
-	}
 
 	@ModifyExpressionValue(
 		method = "submit(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;ILnet/minecraft/client/renderer/entity/state/AvatarRenderState;FF)V",
@@ -61,11 +41,10 @@ public class CapeLayerMixin {
 	)
 	public ClientAsset.Texture frozenLib$useFrozenLibCape(
 		ClientAsset.Texture original,
-		@Share("frozenLib$newCapeTexture") LocalRef<ClientAsset.Texture> newCapeAssetRef
+		@Local(argsOnly = true) AvatarRenderState state
 	) {
-		final ClientAsset.Texture newCapeAsset = newCapeAssetRef.get();
+		final ClientAsset.Texture newCapeAsset = state.getData(ClientCapeUtil.CAPE_TEXTURE_DATA_KEY);
 		if (newCapeAsset != null) return newCapeAsset;
 		return original;
 	}
-
 }