+++ src/main/java/net/frozenblock/lib/config/frozenlib_config/FrozenLibConfig.java
@@ -33,7 +33,7 @@ public class FrozenLibConfig {
 	public static final ConfigEntry<Boolean> REMOVE_EXPERIMENTAL_WARNING = CONFIG.entry("removeExperimentalWarning", EntryType.BOOL, false);
 	public static final ConfigEntry<Boolean> WARDEN_SPAWN_TRACKER_COMMAND = CONFIG.entry("wardenSpawnTrackerCommand", EntryType.BOOL, false);
 	public static final ConfigEntry<Boolean> FILE_TRANSFER_SERVER = CONFIG.entry("fileTransferServer", EntryType.BOOL, true);
-	public static final ConfigEntry<Boolean> FILE_TRANSFER_CLIENT = CONFIG.entry("fileTransferClient", EntryType.BOOL, true);
+	public static final ConfigEntry<Boolean> FILE_TRANSFER_CLIENT = CONFIG.unsyncableEntry("fileTransferClient", EntryType.BOOL, true);
 	public static final ConfigEntry<FrozenLibModResourcePackApi.PackDownloadSetting> PACK_DOWNLOADING = CONFIG.entry(
 		"packDownloading",
 		FrozenLibModResourcePackApi.PackDownloadSetting.ENTRY_TYPE,